MFC 261811,282660,282706: Place VM objects on the object list when created and never remove them.
261811:
Fix function name in KASSERT().

282660:
Place VM objects on the object list when created and never remove them.
This is ok since objects come from a NOFREE zone and allows objects to
be locked while traversing the object list without triggering a LOR.

Ensure that objects on the list are marked DEAD while free or stillborn,
and that they have a refcount of zero.  This required updating most of
the pagers to explicitly mark an object as dead when deallocating it.
(Only the vnode pager did this previously.)

282706:
Satisfy vm_object uma zone destructor requirements after r282660 when
vnode object creation raced.


git-svn-id: https://svn.freebsd.org/base/stable/9@284100 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f