Generalised support for copy-on-write structures shared by threads.
Thread credentials are maintained as follows: each thread has a pointer to
creds and a reference on them. The pointer is compared with proc's creds on
userspace<->kernel boundary and updated if needed.

This patch introduces a counter which can be compared instead, so that more
structures can use this scheme without adding more comparisons on the boundary.


git-svn-id: https://svn.freebsd.org/base/head@284214 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f