MFC r284351: Simplify printaddr(). No functional change.
MFC r284352:
Allow more than one socket entry for a file descriptor. This is needed
for supporting 1-to-many style SCTP sockets. For other sochets, there
is no functional change.

MFC r284353:
Allow more than one local or remote address per socket. This is needed to
support SCTP (and MPTCP in the future). No functional change for existing
protocols.

MFC r284354:
Add support for SCTP.

MFC r284355:
When using -L the code skips a socket if the local or foreign
address is loopback. So it is shown if both are not loopback.
The man page says that it is shown if the local or foreign
address is not loopback. Change the man page to reflect the
code.


git-svn-id: https://svn.freebsd.org/base/stable/10@284441 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f