MFC r284127: witness: don't warn about matrix inconsistencies without holding the mutex
Lock order checking is done without the witness mutex held, so multiple
threads that are racing to establish a new lock order may read matrix
entries that are in an inconsistent state. Don't print a warning in this
case, but instead just redo the check after taking the witness lock.


git-svn-id: https://svn.freebsd.org/base/stable/10@284653 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f