Split the DMAR unit domains and contexts.  Domains carry address space and related data structures.  Contexts attach requests initiators to domains.  There is still 1:1 correspondence between contexts and domains on the running system, since only busdma currently allocates them, using dmar_get_ctx_for_dev().
Large part of the change is formal rename of the ctx to domain, but
patch also reworks the context allocation and free to allow for
independent domain creation.

The helper dmar_move_ctx_to_domain() is introduced for future use, to
reassign request initiator from one domain to another.  The hard issue
which is not yet resolved with the context move is proper handling (or
reserving) RMRR entries in the destination domain as required by ACPI
DMAR table for moved context.

Tested by:	pho
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/head@284869 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f