Limit the number of autoconfigured queues to 8.
This limit was in a previous version of the driver, but it is being re-
added to match the behavior of previous versions of 10. It prevents the
driver from using too many MSI-X vectors on systems with a large number of
logical CPU cores.

Thanks to <jwd@slowblink.com> for bringing up this issue.

Approved by:	jfv (mentor)


git-svn-id: https://svn.freebsd.org/base/stable/10@284876 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f