Latest clang uses openat(2).
If the pathname is absolute or dirfd is AT_FDCWD we can
handle it exactly like open(2).
Otherwise we output an A record to indicate that the path of
an open directory needs to be used (earlier in the trace).

Also filemon_pid_check needs to hold proctree_lock
and use proc_realparent()

Differential Revision:  D2810
Reviewed by: jhb
Approved by: re


git-svn-id: https://svn.freebsd.org/base/stable/10@285112 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f