Add the GEOM_LABEL option to the BEAGLEBONE, GUMSTIX, and RPI-B kernel configuration files, resolving an issue where the UFS and MSDOSFS partitions would not mount as set in fstab(5).
This is a direct commit to stable/10, as the GEOM_LABEL option
is handled differently in head for arm/armv6.  The WANDBOARD
and PANDABOARD already have this kernel option entry via the IMX6
kernel configuration file, so do not need to be changed.

Approved by:	re (kib)
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/10@285132 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f