MFC: r284931
Set the initial system time to a sane (as in: not end of 21st century)
value when booting on a PC with CMOS clock set to a year before 2000.

This uses 1980 (instead of 1970 as in the initial patch) as pivot year as
suggested by imp in the PR followup.

PR:		195703
Submitted by:	cs@soi.spb.ru
Reviewed by:	imp


git-svn-id: https://svn.freebsd.org/base/stable/9@285447 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f