MFC r285190:
Fix if_loop so bpfwrite() can use it regardless of the state of
bd_hdrcmplt.  As if_loop does not use link-level headers, its behavior
when used by bpfwrite() should be the same regardless of the state of
bd_hdrcmplt.  Without this change, libpcap (and other BPF users that
work like it) fail when writing to loopback interfaces.

Approved by: re


git-svn-id: https://svn.freebsd.org/base/stable/10@285605 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f